// src/types/jobs/jobDetail.ts
export interface JobDetailData {
  id: string;
  title: string;
  location: string;
  department: string;
  workType: string;
  aboutUs: string;
  aboutRole: string;
  keyResponsibilities: string[];
  qualifications: string[];
  skillsForSuccess: string[];
  whyJoinUs: string;
  compensation: string;
  
}